Adds simple Machine Learning models and Deep Learning models in one directory.
